Housekeeping: move simulator helper classes to a Python package.
Cleans up `conftest.py` and opens the door for future reuse of simulator classes.

- modified:   test/conftest.py
  - Moved simulator classes out.
  - Import them instead.

- new file:   test/resources/helpers/simulator/pyproject.toml
  - Package config file enables us to install the simulator package with pip, i.e., `pip install -e .`

- new file:   test/resources/helpers/simulator/src/simulator/__init__.py
  - Identifies `simulator` as a package.

- new file:   test/resources/helpers/simulator/src/simulator/questa.py
- new file:   test/resources/helpers/simulator/src/simulator/xcelium.py
  - Moved concrete simulator subclasses into their own modules.
  - It's cleaner to separate the simulators.

- new file:   test/resources/helpers/simulator/src/simulator/simulator.py
  - Moved simulator base class into its own module.
  - This module has the same name as the package, but that doesn't mean the base class module is the entry point for the package.
